moved tagger service to different directory
